Removed check for misconfigured Voter with constructor argument
This check was inconsequent (no other form of voter description had a
similar check) and not very helpful (due to the general error message).

For now, the removal simplifies the code. If necessary, one could add
other checks and improve the error messages later on, if the
instructions in the readme do not suffice.