Read attributes as provided by the event object, and pass on attributes to other event handlers (Case 177753) (#15)
This adapts
https://github.com/webfactory/WebfactoryWfdMetaBundle/pull/53:

> This PR uses the new API introduced in
https://github.com/symfony/symfony/pull/46001 to read controller
attributes through the `ControllerEvent`, and to make them available to
other event handlers when replacing the controller.
>
> This is necessary when using the `Send304IfNotModified` attribute in
combination with `\Symfony\Component\HttpKernel\Attribute\Cache`.
Without this change,
`\Symfony\Component\HttpKernel\EventListener\CacheAttributeListener`
will not set `Cache` headers accordingly. The result is that you may get
`304 Not Modified` responses on conditional requests with
`If-Modified-Since`, but these are treated as `stale/cache` only in the
HttpCache and have a `Cache-Control: must-revalidate, private` header.

And adds some tests.